refactor(cocache): rename CacheSourceResolver to CacheSourceFactory
- Rename CacheSourceResolver interface to CacheSourceFactory
- Update related implementations and imports
- Adjust package structure for better organization
